+++ gammapy/cube/make.py
@@ -8,7 +8,7 @@ from gammapy.modeling.models import BackgroundModel
 from .background import make_map_background_irf
 from .edisp_map import make_edisp_map
 from .exposure import make_map_exposure_true_energy
-from .fit import MapDataset, MapDatasetOnOff
+from gammapy.datasets.fit import MapDataset, MapDatasetOnOff
 from .psf_map import make_psf_map
 
 __all__ = ["MapDatasetMaker", "SafeMaskMaker"]
