+++ gammapy/datasets/__init__.py
@@ -1,2 +1,3 @@
 from .core import *
 from .spectrum import *
+from .fit import *
