+++ gammapy/datasets/tests/test_fit.py
@@ -7,8 +7,6 @@ from astropy.coordinates import SkyCoord
 from regions import CircleSkyRegion
 from gammapy.cube import (
     EDispMap,
-    MapDataset,
-    MapDatasetOnOff,
     PSFMap,
     make_map_exposure_true_energy,
 )
@@ -16,7 +14,7 @@ from gammapy.data import GTI
 from gammapy.irf import EffectiveAreaTable2D, EnergyDependentMultiGaussPSF
 from gammapy.maps import Map, MapAxis, WcsGeom, WcsNDMap
 from gammapy.modeling import Fit
-from gammapy.datasets import Datasets
+from gammapy.datasets import Datasets, MapDataset, MapDatasetOnOff
 from gammapy.modeling.models import (
     BackgroundModel,
     GaussianSpatialModel,
