+++ gammapy/maps/tests/test_wcsnd.py
@@ -9,7 +9,7 @@ from astropy.io import fits
 from astropy.table import Table
 from regions import CircleSkyRegion
 from gammapy.cube import PSFKernel
-from gammapy.cube.fit import MapEvaluator
+from gammapy.datasets.fit import MapEvaluator
 from gammapy.irf import EnergyDependentMultiGaussPSF
 from gammapy.maps import Map, MapAxis, MapCoord, WcsGeom, WcsNDMap
 from gammapy.maps.utils import fill_poisson
